requirements.txt
gradio>=4.0.0
langchain>=0.0.338
langchain-core>=0.1.0
langgraph>=0.0.17
langchain-openai>=0.0.5
openai>=1.6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
json-repair>=0.4.0
requests>=2.31.0
tavily-python>=0.3.1
tenacity>=8.2.0
typing-extensions>=4.8.0
urllib3>=2.0.0 